metadata:  Fix sign extension in get_uint64_le() routine

+++ lib/rbcodec/metadata/metadata_common.c
@@ -149,7 +149,7 @@ uint64_t get_uint64_le(void* buf)
 {
     unsigned char* p = (unsigned char*) buf;
 
-    return p[0] | (p[1] << 8)  | (p[2] << 16) | (p[3] << 24) | ((uint64_t)p[4] << 32) |
+    return ((uint64_t)p[0]) | ((uint64_t)p[1] << 8)  | ((uint64_t)p[2] << 16) | ((uint64_t)p[3] << 24) | ((uint64_t)p[4] << 32) |
                   ((uint64_t)p[5] << 40) | ((uint64_t)p[6] << 48) | ((uint64_t)p[7] << 56);
 }
